refactor: update imports from @/lib/data to @/constants
Moved data files to constants and updated import paths
