+++ components/ui/hero-stats-marquee.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { getEventStats } from "@/lib/data/events";
+import { getEventStats } from "@/constants/events";
 
 export function HeroStatsMarquee() {
   const stats = getEventStats();
