+++ constants/sponsors.ts
